redid light
completely re-worked how light is calculated for each pixel. produces much cleaner and more even light and shadows, including colored lights (which should no longer have that dark ring around the edges). was messing around with the ceiling being a light, so some of the code from that to make it work is still left in there (will be cleaned up later). lots of messy comments and useless chunks of code now, which needs to be cleaned up.